pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-sync"
version = "0.1.0"
description = "One soul, many bodies — sync AI agent identity, memory, and knowledge across platforms"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "jasonjcwu" },
]
keywords = ["ai", "agent", "harness", "sync", "openclaw", "claude-code", "codex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "rich>=13.0",
    "pathspec>=0.12",
]

[project.scripts]
agent-sync = "agent_sync.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
